requirements.txt
aiohttp
google-genai
trafilatura
pdfplumber
docx2txt
scikit-learn
python-dotenv
numpy
streamlit
